% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiDefinitions.R
\name{Dimension}
\alias{Dimension}
\title{The class contains dimension description and dimension items}
\usage{
Dimension(data)
}
\description{
The class contains dimension description and dimension items
}
