% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiDefinitions.R
\name{DimensionModel}
\alias{DimensionModel}
\title{This module contains metadata definitions for Knoema API
The class contains dimension description}
\usage{
DimensionModel(data)
}
\description{
This module contains metadata definitions for Knoema API
The class contains dimension description
}
