% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonFunctions.R
\name{isequal_strings_ignorecase}
\alias{isequal_strings_ignorecase}
\title{Helper functions
The function compares strings ignoring case}
\usage{
isequal_strings_ignorecase(first, second)
}
\description{
Helper functions
The function compares strings ignoring case
}
