% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonFunctions.R
\name{IsEqualStringsIgnoreCase}
\alias{IsEqualStringsIgnoreCase}
\title{Helper functions
The function compares strings ignoring case}
\usage{
IsEqualStringsIgnoreCase(first, second)
}
\description{
Helper functions
The function compares strings ignoring case
}
