% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ApiDefinitions.R
\name{PivotTimeItem}
\alias{PivotTimeItem}
\title{The class contains pivot request item}
\usage{
PivotTimeItem(dimensionid = NULL, members = NULL, uimode = NULL)
}
\description{
The class contains pivot request item
}
