% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataReader.R
\name{DataReader}
\alias{DataReader}
\title{This class read data from Knoema and transform it to ts,xts or zoo}
\usage{
DataReader(client, dataset, selection)
}
\description{
This class read data from Knoema and transform it to ts,xts or zoo
}
