% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knoema.R
\name{knoema.get}
\alias{knoema.get}
\title{knoema package!
This package works with datasets from knoema.com}
\usage{
knoema.get(datasetId, selection, type = "ts", host = "", app_id = "",
  app_secret = "")
}
\arguments{
\item{datasetId}{is Dataset's ID specified as a string}

\item{selection}{is list where all the dimensions of the dataset are listed and a selection from them}

\item{type}{is optional. By default equal "ts". Other supported variants are "xts" and "zoo"}

\item{client}{is optional. By default use public user for knoema.com
By default the package allows you to work only with public datasets from the site knoema.com.
If you want work with private datasets or from other hosts, you need create ApiClient
client = ApiClient("host","some_app_id","some_app_secret") and then use this client in funcion knoema.get
You can get parameters app_id and app_secret after registering on the site knoema.com, in the section "My profile - Apps - create new" (or use existing applications)}
}
\description{
knoema package!
This package works with datasets from knoema.com
}
\examples{
knoema.get('IMFWEO2017Apr',list(country="512;914",subject ="NGDP_RPCH"))
knoema.get('IMFWEO2017Apr',list(country="512;914",subject ="NGDP_RPCH"),type="xts")
knoema.get('IMFWEO2017Apr',list(country="512;914",subject ="NGDP_RPCH"),type="zoo", app_id="some app_id", app_secret="some app_secret")
}
